updated structure to have components in their own files, cleaner to read now
